+++ src/core/router/engines/browser-history.ts
@@ -26,6 +26,18 @@ import type { Router, Route, HistoryClearFilter } from 'core/router/interface';
 export const
 	$$ = symbolGenerator();
 
+const
+	isIFrame = location !== parent.location;
+
+/**
+ * This code is needed to fix a bug with the History API router engine when backing to the
+ * first history item doesn’t emit a popstate event in Safari if the script is running within an iframe
+ * @see https://github.com/V4Fire/Client/issues/717
+ */
+if (isIFrame && (browser.is.Safari !== false || browser.is.iOS !== false)) {
+	history.pushState({}, '', location.href);
+}
+
 /**
  * This flag is needed to get rid of a redundant router transition when restoring the page from BFCache in safari
  * @see https://github.com/V4Fire/Client/issues/552