+++ tests/helpers/dom/interface.ts
@@ -8,13 +8,13 @@
 
 export interface WaitForElOptions {
 	/**
-	 * The delay between trying to find an element on a page
+	 * The delay between trying to find an element
 	 * @default `100`
 	 */
 	sleep?: number;
 
 	/**
-	 * Time after which the function stops trying to find an element on a page and returns `undefined`
+	 * Time after which the function stops trying to find an element and returns `undefined`
 	 * @default `3500`
 	 */
 	timeout?: number;
@@ -22,8 +22,8 @@ export interface WaitForElOptions {
 	/**
 	 * Event to wait
 	 *
-	 *  * `mount` – element attached to a page
-	 *  * `unmount` – element detached from a page
+	 *  1. `mount` – element attached to a page;
+	 *  2. `unmount` – element detached from a page.
 	 *
 	 * @default `mount`
 	 */